chore: update tests to playwright
