+++ tests/e2e/core/version.spec.ts
